add `__func__` and `__self__` to WeakMethod and deref WeakMethod in `func_to_source_code`
